+++ tests/test-config-file/webpack.custom.config.js
@@ -1,0 +1,10 @@
+const HtmlWebpackPlugin = require('html-webpack-plugin');
+
+module.exports = {
+  entry: './app.js',
+  output: {
+     path: './custom',
+     filename: 'app.custom.bundle.js'
+  },
+  plugins: [new HtmlWebpackPlugin()],
+ };